Update and rename install.sh to install_Agama.sh
Make the script aware of which directory is this script installed on, for in the case it is executed in a clone of the repository avoid cloning it again as a subdirectory, because it leads to duplicate files and confusion to the developers. I don't know if the single branch cloning is to save disk space, but if it is done on a subdirectory of the first-made clone, we are not saving space; It would be better to do a single clone, and if desired to do it with a single branch, but to do it from outside of an agama clone, or if the install script is done  from a clone, avoud re-cloning to avoid data confusion and to save disk space.